FSDEX-2247 Blueshift Amm investigation - Updated data models

+++ src/main/java/fi/blueshift/lib/domain/dto/TokenPairTradingConfigDto.java
@@ -38,7 +38,7 @@ public class TokenPairTradingConfigDto implements Serializable {
         Integer levelIndex;
         Double minProfitPercentage;
         Double maxProfitPercentage;
-        BigDecimal liquidityAmount;
+        Double liquidityAmount;
         Integer ordersCountForLevel;
     }
 }
